add feature mixin
